feat: Criando Models turmas, arquitetura MVC
